refactor(responses): split responses.py by domain into a package
responses.py held 35 GraphQL envelope models across every domain. Split
into a responses/ package with one file per domain (auth, installation,
alarm, activity, lock, camera, sentinel, errors) plus a _base.py for
the shared inner fragments (_ResMsg, _ResMsgRef, PanelError,
_OperationResult). responses/__init__.py re-exports everything so
existing imports keep working unchanged. Pure file-level
reorganisation — no behaviour change. All tests pass.
